Formations/OC-MLE/Dog-Breed-Detector/applications/flask_dog_breed_detector/requirements.txt
tensorflow==2.4.0
numpy
Flask
flask_monitoringdashboard
opencv_python
matplotlib
tensorflow_hub
six
Pillow
schedule
gunicorn
